requirements.txt
absl-py==1.4.0
addict==2.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
antlr4-python3-runtime==4.9.3
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
basicsr==1.4.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
clean-fid==0.1.35
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
einops==0.6.1
facexlib==0.3.0
fastapi==0.95.2
ffmpy==0.3.0
filelock==3.12.0
filterpy==1.4.5
font-roboto==0.0.1
fonts==0.0.3
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.3
gfpgan==1.3.8
google-auth==2.19.1
google-auth-oauthlib==1.0.0
gradio==3.33.0
gradio_client==0.2.5
grpcio==1.54.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
idna==3.4
imageio==2.30.0
Jinja2==3.1.2
jsonmerge==1.9.0
jsonschema==4.17.3
kiwisolver==1.4.4
kornia==0.6.12
lark==1.1.5
lazy_loader==0.2
lightning-utilities==0.8.0
linkify-it-py==2.0.2
llvmlite==0.40.0
lmdb==1.4.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.57.0
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
orjson==3.9.0
packaging==23.1
pandas==2.0.2
piexif==1.1.3
Pillow==9.5.0
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.8
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytorch-lightning==2.0.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
realesrgan==0.3.0
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
resize-right==0.0.2
rsa==4.9
scikit-image==0.20.0
scipy==1.10.1
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tb-nightly==2.14.0a20230601
tensorboard-data-server==0.7.0
tifffile==2023.4.12
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchdiffeq==0.2.3
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.2
typing_extensions==4.6.3
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==1.26.16
uvicorn==0.22.0
websockets==11.0.3
Werkzeug==2.3.4
yapf==0.33.0
yarl==1.9.2
